requirements.txt
fastapi==0.111.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.69
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
enum34==1.1.10
pytest==8.0.2
httpx==0.27.0
psycopg2-binary==2.9.9
